app/src/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
